moved the image render in the mysnapps page to the utils folder as ImageRender.js
